Distinguish between One and All methods when using get and remove
Because the 'text' argument is optional, there is no way to distinguish
between getOne/getAll and removeOne/removeAll when using the shorthand
get/remove syntax. So I've set it so that you need to explicitly make
the second argument `false` if you want to use the 'all' functionality.

So basically, `tinyQuery.get('/',false) === tinyQuery.getAll('/')`.